split out cli interface methods into a mixin for reuse
