+++ awslambda/python_requirements/dependency_managers/_poetry.py
@@ -99,7 +99,7 @@ class Poetry(DependencyManager):
                     without_hashes=without_hashes,
                 )
             )
-            requirements_txt = self.root_directory / output.name
+            requirements_txt = self.cwd / output.name
             if requirements_txt.is_file():
                 output.parent.mkdir(exist_ok=True, parents=True)
                 return requirements_txt.rename(output)
